refactor : 도메인 내에서 패키지 분리
- ladder, result, user 패키지로 분리 및 각각 Creator클래스 생성
