+++ src/main/java/domain/ladder/LadderCreator.java
@@ -1,0 +1,8 @@
+package domain.ladder;
+
+public class LadderCreator {
+    public static Ladder createLadder(int ladderHeight, int ladderWidth) {
+        return new Ladder(ladderHeight, ladderWidth);
+    }
+
+}
