+++ src/main/java/domain/ladder/Line.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import util.RandomUtil;
