+++ src/test/java/domain/LineTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.ladder.Line;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
